Created a common base template.
So inside this there a lot of blocks in which we are gonna include the
content. The content inside is the default content if we don’t include
anything else inside it in our template.
Blocks I have created:
desc - For meta description;
titlename - For the title;
head - For extra head content such as adding specific stylesheats for
specific pages;
style - For adding css in page;
headerclass - default is fixed header, you can change it to a normal
header;
content - For the page content;
afterfooter - For content after the footer;
js - For attaching javascript files;